requirements.txt
# Core dependencies for FastAPI inference API (Python 3.11+)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
Pillow>=11.0.0
PyYAML>=6.0.2
python-multipart>=0.0.9
pydantic>=2.9.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
numpy>=2.0.0

# PyTorch CPU-only (muito mais leve para deploy em nuvem)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.5.0
torchvision>=0.20.0

